+++ src/lemming/__main__.py
@@ -17,13 +17,14 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 # SPDX-License-Identifier: GPL-3.0-or-later
+import dataclasses
 import pathlib
 import sys
 import time
 from typing import List, Optional
 
 import typer
-from typing_extensions import Annotated, Self, TypedDict
+from typing_extensions import Annotated, Self
 
 from . import __version__, logger
 from .config import (
@@ -56,13 +57,33 @@ exit $ret_code
 app = typer.Typer(no_args_is_help=True)
 
 
-class Settings(TypedDict):
+@dataclasses.dataclass(frozen=True)
+class Settings:
+    """
+    The settings.
+
+    Args:
+        what_to_quiet (WhatToQuiet): What to quiet.
+        config (Config): The configuration.
+        only (Optional[List[str]]): Only run these formatters/linters. If
+        None, run all.
+    """
+
     what_to_quiet: WhatToQuiet
     config: Config
+    only: Optional[List[str]]
+
+    def should_run(self, name: str) -> bool:
+        if self.only is None:
+            return True
+        return name in self.only
 
 
 class Timer:
+    """A timer."""
+
     def __init__(self) -> None:
+        """Create a Timer instance."""
         self.start = None
         self.time = None
 
@@ -78,16 +99,30 @@ class Timer:
 def run_linter(
     linter: Linter, paths: List[pathlib.Path], settings: Settings
 ) -> bool:
+    """
+    Run `linter`.
+
+    Args:
+        linter (Linter): The linter to run.
+        paths (List[pathlib.Path]): The paths to lint.
+        settings (Settings): The settings.
+
+    Raises:
+        typer.Exit: If the linter failed and fail fast is True.
+
+    Returns:
+        bool: `exit_status == 0`
+    """
     with logger.ctxmgr:
         with Timer() as linter_timer:
-            success = linter.run(paths, settings["what_to_quiet"])
+            success = linter.run(paths, settings.what_to_quiet)
             if not success:
                 logger.error(
-                    f"Could not run linter {linter.packages}!"
+                    f"Could not run linter {linter.name}!"
                     " Please see the linter's output for more"
                     " details."
                 )
-                if settings["config"].fail_fast:
+                if settings.config.fail_fast:
                     raise typer.Exit(1)
                 return False
         logger.info(f"Ran linter in {linter_timer.time} seconds")
@@ -95,12 +130,25 @@ def run_linter(
 
 
 def linter_first(paths: List[pathlib.Path], settings: Settings) -> bool:
+    """
+    Run the first linters.
+
+    Args:
+        paths (List[pathlib.Path]): The paths to lint.
+        settings (Settings): The settings.
+
+    Returns:
+        bool: Whether or not all linters successfully ran.
+    """
     logger.info("Running first linters")
     return_value = True
     with logger.ctxmgr:
         with Timer() as linters_timer:
-            for linter in settings["config"].get_first_linters():
-                logger.info(f"Running first linter {linter.packages}")
+            for linter in settings.config.get_first_linters():
+                if not settings.should_run(linter.name):
+                    logger.info(f"Skipping first linter {linter.name}")
+                    continue
+                logger.info(f"Running first linter {linter.name}")
                 success = run_linter(linter, paths, settings)
                 if not success:
                     return_value = False
@@ -114,18 +162,31 @@ def run_formatter(
     format_: bool,
     settings: Settings,
 ) -> bool:
+    """
+    Run `formatter`.
+
+    Args:
+        formatter (Formatter): The formatter to run.
+        paths (List[pathlib.Path]): The paths to format/check.
+        format_ (bool): Whether or not to format or check.
+        settings (Settings): The settings.
+
+    Raises:
+        typer.Exit: If the formatter failed and fail fast is True.
+
+    Returns:
+        bool: `exit_status == 0`
+    """
     with logger.ctxmgr:
         with Timer() as formatter_timer:
             if format_:
-                success = formatter.run_format(
-                    paths, settings["what_to_quiet"]
-                )
+                success = formatter.run_format(paths, settings.what_to_quiet)
             else:
-                success = formatter.run_check(paths, settings["what_to_quiet"])
+                success = formatter.run_check(paths, settings.what_to_quiet)
 
             if not success:
-                logger.error(f"Could not run formatter {formatter.packages}!")
-                if settings["config"].fail_fast:
+                logger.error(f"Could not run formatter {formatter.name}!")
+                if settings.config.fail_fast:
                     raise typer.Exit(1)
                 return False
         logger.info(f"Ran formatter in {formatter_timer.time} seconds")
@@ -135,12 +196,26 @@ def run_formatter(
 def formatter(
     format_: bool, paths: List[pathlib.Path], settings: Settings
 ) -> bool:
+    """
+    Run all formatters.
+
+    Args:
+        format_ (bool): Whether or not to format or check.
+        paths (List[pathlib.Path]): The paths to format/check.
+        settings (Settings): The settings.
+
+    Returns:
+        bool: Whether or not all formatters successfully ran.
+    """
     logger.info("Running formatters")
     return_value = True
     with logger.ctxmgr:
         with Timer() as formatters_timer:
-            for formatter in settings["config"].formatters:
-                logger.info(f"Running formatter {formatter.packages}")
+            for formatter in settings.config.formatters:
+                if not settings.should_run(formatter.name):
+                    logger.info(f"Skipping formatter {formatter.name}")
+                    continue
+                logger.info(f"Running formatter {formatter.name}")
                 success = run_formatter(formatter, paths, format_, settings)
                 if not success:
                     return_value = False
@@ -149,12 +224,25 @@ def formatter(
 
 
 def linter_other(paths: List[pathlib.Path], settings: Settings) -> bool:
+    """
+    Run all other linters.
+
+    Args:
+        paths (List[pathlib.Path]): The paths to lint.
+        settings (Settings): The settings.
+
+    Returns:
+        bool: Whether or not all other linters successfully ran.
+    """
     logger.info("Running other linters")
     return_value = True
     with logger.ctxmgr:
         with Timer() as linters_timer:
-            for linter in settings["config"].get_other_linters():
-                logger.info(f"Running other linter {linter.packages}")
+            for linter in settings.config.get_other_linters():
+                if not settings.should_run(linter.name):
+                    logger.info(f"Skipping other linter {linter.name}")
+                    continue
+                logger.info(f"Running other linter {linter.name}")
                 success = run_linter(linter, paths, settings)
                 if not success:
                     return_value = False
@@ -163,6 +251,17 @@ def linter_other(paths: List[pathlib.Path], settings: Settings) -> bool:
 
 
 def run(paths: List[pathlib.Path], format_: bool, settings: Settings) -> None:
+    """
+    Run all linters and formatters.
+
+    Args:
+        paths (List[pathlib.Path]): The paths to check.
+        format_ (bool): Whether or not to format or check.
+        settings (Settings): The settings.
+
+    Raises:
+        typer.Exit: If any formatter or linter failed and fail fast is False.
+    """
     with Timer() as all_timer:
         success = True
         if linter_first(paths, settings) is False:
@@ -185,21 +284,52 @@ def run(paths: List[pathlib.Path], format_: bool, settings: Settings) -> None:
 
 
 def quiet_callback(value: bool) -> bool:
+    """
+    Increase the logger's threshold by 10 if `value`.
+
+    Args:
+        value (bool): Value.
+
+    Returns:
+        bool: `value`
+    """
     if value:
         logger.threshold += 10
     return value
 
 
 def verbose_callback(value: bool) -> bool:
+    """
+    Decrease the logger's threshold by 10 if `value`.
+
+    Args:
+        value (bool): Value.
+
+    Returns:
+        bool: `value`
+    """
     if value:
         logger.threshold -= 10
     return value
 
 
-def version_callback(value: bool) -> None:
+def version_callback(value: bool) -> bool:
+    """
+    Print the version and exit if `value`.
+
+    Raises:
+        typer.Exit: if `value`
+
+    Args:
+        value (bool): Value.
+
+    Returns:
+        bool: `value`
+    """
     if value:
         print(__version__)
         raise typer.Exit(0)
+    return value
 
 
 def both(
@@ -208,7 +338,30 @@ def both(
     verbose: bool,
     quiet: bool,
     config: Optional[pathlib.Path],
+    only: Optional[List[str]],
 ) -> Settings:
+    """
+    Create the settings, and do some stuff with the given options.
+
+    Args:
+        quiet_commands (bool): Whether or not to stop the commands from
+        writing to stdout and stderr.
+        quiet_pip (bool): Whether ot not to stop pip from writing to stdout
+        and stderr
+        verbose (bool): Decrease the loggers threshold by 10 (done by the
+        callback).
+        quiet (bool): Increase the loggers threshold by 10 (done by the
+        callback).
+        config (Optional[pathlib.Path]): The configuration to use or None.
+        only (Optional[List[str]]): Only run these formatters/linters. If
+        None, run all.
+
+    Raises:
+        typer.Exit: If both verbose and quiet is passed.
+
+    Returns:
+        Settings: The settings.
+    """
     if verbose and quiet:
         logger.critical("Verbose and quiet are mutually exclusive!")
         raise typer.Exit(2)
@@ -223,6 +376,7 @@ def both(
     return Settings(
         what_to_quiet=WhatToQuiet(commands=quiet_commands, pip=quiet_pip),
         config=config_,
+        only=only,
     )
 
 
@@ -274,9 +428,20 @@ def format_(
             exists=True, dir_okay=False, help="The config file to use."
         ),
     ] = None,
+    only: Annotated[
+        Optional[List[str]],
+        typer.Option(
+            help="Only run these formatters/linters (may be passed multiple"
+            " times)"
+        ),
+    ] = None,
 ) -> None:
     """Format your code and run linters."""
-    run(paths, True, both(quiet_commands, quiet_pip, verbose, quiet, config))
+    run(
+        paths,
+        True,
+        both(quiet_commands, quiet_pip, verbose, quiet, config, only),
+    )
 
 
 @app.command()
@@ -324,15 +489,39 @@ def check(
     config: Annotated[
         Optional[pathlib.Path],
         typer.Option(
-            exists=True, dir_okay=False, help="The config file to use."
+            exists=True, dir_okay=False, help="The config file to use"
+        ),
+    ] = None,
+    only: Annotated[
+        Optional[List[str]],
+        typer.Option(
+            help="Only run these formatters/linters (may be passed multiple"
+            " times)"
         ),
     ] = None,
 ) -> None:
     """Check the formatting of your code and run linters."""
-    run(paths, False, both(quiet_commands, quiet_pip, verbose, quiet, config))
+    run(
+        paths,
+        False,
+        both(quiet_commands, quiet_pip, verbose, quiet, config, only),
+    )
 
 
 def install_pre_commit(git_repository: pathlib.Path) -> None:
+    """
+    Install the pre-commit hook script to
+    `git_repository/.git/hooks/pre-commit`
+
+    Args:
+        git_repository (pathlib.Path): The root directory of the git
+        repository.
+
+    Raises:
+        typer.Exit: If `git_repository` doesn't exist
+        typer.Exit: If `git_repository/.git` doesn't exist
+        typer.Exit: If writing to the pre-commit file failed.
+    """
     if not git_repository.exists():
         logger.critical(f"Directory {git_repository} does not exist!")
         raise typer.Exit(1)
@@ -411,6 +600,7 @@ def callback(
 
 
 def main() -> None:
+    """Main."""
     return app()